migrated player setup from gameSetup to gameLoop.
