+++ DragonTreasure/src/DragonTreasure.java
@@ -13,14 +13,10 @@ public class DragonTreasure{
     public static void main(String[] args){
         // logic to setup the game.
         GameSetup setupDungeon = new GameSetup();
-        // prompt user for username and welcome player to game.
-        setupDungeon.promptUserName();
-
         // get data to create gameloop from setup.
         Room[] dungeonRooms = setupDungeon.getRooms();
-        Player newPlayer = setupDungeon.getPlayer();
         // create loop.
-        GameLoop loop = new GameLoop(newPlayer, dungeonRooms);
+        GameLoop loop = new GameLoop(dungeonRooms);
         // start a gameloop.
         loop.start();
         // NOTE: we need to sleep a little here otherwise game will just terminate immediately.
